refactor: extract image output transformation
